Sync the fork

+++ modules/activiti-engine/src/test/java/org/activiti/examples/task/StandaloneTaskTest.java
@@ -17,8 +17,8 @@ import java.util.List;
 import java.util.Map;
 
 import org.activiti.engine.ActivitiOptimisticLockingException;
-import org.activiti.engine.history.HistoricDetail;
 import org.activiti.engine.history.HistoricVariableInstance;
+import org.activiti.engine.impl.history.HistoryLevel;
 import org.activiti.engine.impl.persistence.entity.TaskEntity;
 import org.activiti.engine.impl.test.PluggableActivitiTestCase;
 import org.activiti.engine.task.Task;
@@ -107,7 +107,7 @@ public class StandaloneTaskTest extends PluggableActivitiTestCase {
     taskService.deleteTask(taskId, true);
   }
   
-  // See http://jira.codehaus.org/browse/ACT-1290 
+  // See https://activiti.atlassian.net/browse/ACT-1290
   public void testRevisionUpdatedOnSave() {
     Task task = taskService.newTask();
     taskService.saveTask(task);
@@ -124,7 +124,7 @@ public class StandaloneTaskTest extends PluggableActivitiTestCase {
     taskService.deleteTask(task.getId(), true);
   }
 
-  // See http://jira.codehaus.org/browse/ACT-1290 
+  // See https://activiti.atlassian.net/browse/ACT-1290
   public void testRevisionUpdatedOnSaveWhenFetchedUsingQuery() {
     Task task = taskService.newTask();
     taskService.saveTask(task);
@@ -146,31 +146,32 @@ public class StandaloneTaskTest extends PluggableActivitiTestCase {
   }
   
   public void testHistoricVariableOkOnUpdate() {
-  		//1. create a task
+    if (processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.AUDIT)) {
+  		// 1. create a task
   		Task task = taskService.newTask();
   		task.setName("test execution");
   		task.setOwner("josOwner");
   		task.setAssignee("JosAssignee");
   		taskService.saveTask(task);
   		 
-  		//2. set task variables
+  		// 2. set task variables
   		Map<String, Object> taskVariables = new HashMap<String, Object>();
   		taskVariables.put("finishedAmount", 0);
   		taskService.setVariables(task.getId(), taskVariables);
   		 
-  		 
-  		//3. complete this task with a new variable
+  		// 3. complete this task with a new variable
   		Map<String, Object> finishVariables = new HashMap<String, Object>();
   		finishVariables.put("finishedAmount", 40);
   		taskService.complete(task.getId(), finishVariables);
   		 
-  		//4. get completed variable
+  		// 4. get completed variable
   		List<HistoricVariableInstance> hisVarList = historyService.createHistoricVariableInstanceQuery().taskId(task.getId()).list();
   		assertEquals(1, hisVarList.size());
   		assertEquals(40, hisVarList.get(0).getValue());
   		
   		// Cleanup
   		historyService.deleteHistoricTaskInstance(task.getId());
-  	}
+    }
+	}
 
 }
